@@ -1,5 +1,4 @@
-import React from "react";
-import HighlightIcon from "@material-ui/icons/Highlight";
+import HighlightIcon from "@mui/icons-material/Highlight";
 
 function Header() {
   return (
