@@ -1,5 +1,4 @@
-import React from "react";
-import DeleteIcon from "@material-ui/icons/Delete";
+import DeleteIcon from "@mui/icons-material/Delete";
 
 function Note(props) {
   function handleClick() {
